⚡FreeText uses built-in Wordcloud vis
