+++ response-items/freetext/package.json
@@ -54,7 +54,6 @@
     "mathjs": "~7",
     "prop-types": "~15",
     "react": "~17",
-    "react-dom": "~17",
-    "react-wordcloud": "~1"
+    "react-dom": "~17"
   }
 }
